chore: forge-lint Batch 5.1 — 5 redundant cast removals + 3 reclassified suppressions (#264)
Batch 5.1 — 8 sites from the CSV-flagged REDUNDANT-REMOVE bucket. Refined triage reclassified 3 of them to REVIEW-CONTEXT (the API3 int224→int256→uint256 widen + the LibInteractionRewards signed-delta math) — those got inline rationale suppressions instead. Net: 5 redundant casts removed + 3 suppressed = 8 warnings cleared.

Codex pass 1 clean: "Bravo." CI all green.

+++ contracts/test/DepthTieredLtv.t.sol
@@ -145,8 +145,8 @@ contract DepthTieredLtv is Test {
         // `_liquidityTier._tryGetAssetPriceView`) for both legs — pin
         // both to $2,000 / 8-dec so the value-balance guard passes for a
         // `sqrtP = 2⁹⁶` (1:1) pool with equal-decimal tokens.
-        _mockAssetPrice(mockAsset, uint256(uint256(PRICE_2000)), 8);
-        _mockAssetPrice(mockWeth, uint256(uint256(PRICE_2000)), 8);
+        _mockAssetPrice(mockAsset, uint256(PRICE_2000), 8);
+        _mockAssetPrice(mockWeth, uint256(PRICE_2000), 8);
 
         // Default: no pool at any probed tier on any factory. Tests
         // opt-in via `_mockPool` for whichever (factory, fee, L) they need.
@@ -208,7 +208,7 @@ contract DepthTieredLtv is Test {
         // too — a tightening vs the legacy `_v3DepthLiquid` metric which
         // had no value-balance guard and would have said Liquid here.
         _mockPool(mockUniFactory, mockAsset, 3000, L_TIER3);
-        _mockAssetPrice(mockAsset, uint256(uint256(PRICE_2000)) * 2, 8); // asset now "worth 2× WETH"
+        _mockAssetPrice(mockAsset, uint256(PRICE_2000) * 2, 8); // asset now "worth 2× WETH"
         assertEq(
             uint256(OracleFacet(address(diamond)).checkLiquidity(mockAsset)),
             uint256(LibVaipakam.LiquidityStatus.Illiquid)
@@ -464,7 +464,7 @@ contract DepthTieredLtv is Test {
         // honoured. (And `_checkLiquidity` still needs an `asset/WETH`
         // pool — keep that one.)
         ERC20Mock altQuote = new ERC20Mock("Alt", "ALT", 18);
-        _mockAssetPrice(address(altQuote), uint256(uint256(PRICE_2000)), 8);
+        _mockAssetPrice(address(altQuote), uint256(PRICE_2000), 8);
         _mockEmptyAtAllPairsTiers(mockUniFactory, mockAsset, address(altQuote));
 
         _mockPool(mockUniFactory, mockAsset, 3000, L_TIER1); // asset/WETH (keeps it Liquid)
